==> Notebooks/3 Data Preprocessing.ipynb <==
# Data Preprocessing Section

import pandas as pd
data = pd.read_csv('cleaned_train.csv')

data.columns

# Display the data types of each column
print(data.dtypes)

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import LabelEncoder

# Create OneHotEncoder Object
onehot = OneHotEncoder(drop = 'first')

# Occupation is the only column that should need one hot encoding
data_onehot = data[['Occupation']]

data_onehot = onehot.fit_transform(data_onehot.values)

# Make sure the new column names look to be correct
onehot.get_feature_names_out()


# Re-construct the result to a DataFrame
data_onehot = pd.DataFrame(data = data_onehot.toarray(), columns = onehot.get_feature_names_out())

# Update the orginal DataFrame data with the newly OneHot encoded data.
data[onehot.get_feature_names_out()] = data_onehot

# Drop the original column
data.drop(['Occupation'], axis = 1, inplace = True)

# Check the columns
data.info()

# Splitting features and target variable
X = data.drop(columns=['Credit_Score'])
y = data['Credit_Score']

# Splitting the dataset into train and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=101)

# Check to make sure the dimensions are still the same
print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

# Define ordinal mappings for the variables that should be label encoded
credit_mix_mapping = {'Good': 2, 'Standard': 1, 'Bad': 0}
payment_of_min_amount_mapping = {'Yes': 2, 'NM': 1, 'No': 0}
payment_behavior_mapping = {
    'Low_spent_Large_value_payments': 5,
    'Low_spent_Medium_value_payments': 4,
    'Low_spent_Small_value_payments': 3,
    'High_spent_Large_value_payments': 2,
    'High_spent_Medium_value_payments': 1,
    'High_spent_Small_value_payments': 0
}

# Perform the mapping on each of the ordinal columns for both training and testing sets
X_train['Credit_Mix'] = X_train['Credit_Mix'].map(credit_mix_mapping)
X_test['Credit_Mix'] = X_test['Credit_Mix'].map(credit_mix_mapping)

X_train['Payment_of_Min_Amount'] = X_train['Payment_of_Min_Amount'].map(payment_of_min_amount_mapping)
X_test['Payment_of_Min_Amount'] = X_test['Payment_of_Min_Amount'].map(payment_of_min_amount_mapping)

X_train['Payment_Behaviour'] = X_train['Payment_Behaviour'].map(payment_behavior_mapping)
X_test['Payment_Behaviour'] = X_test['Payment_Behaviour'].map(payment_behavior_mapping)


# Check one of the columns to make sure it worked
X_train['Payment_Behaviour']

# Check to make sure X_train and y_train have the same dimensions
print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

# Fill any null value swith 0.0 (might no longer need this part)
X_train = X_train.fillna(0.0)
X_test = X_test.fillna(0.0)

# Check to make sure X_train and y_train have the same dimensions, which they do not so that is a problem 
print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

# Import the necessary things
from sklearn.preprocessing import RobustScaler, MinMaxScaler
from itertools import chain

# Define the numeric columns that will be in the models
numeric_columns = ['Age', 'Annual_Income', 'Monthly_Inhand_Salary', 'Num_Bank_Accounts', 'Num_Credit_Card',
                   'Interest_Rate', 'Num_of_Loan', 'Delay_from_due_date', 'Num_of_Delayed_Payment',
                   'Changed_Credit_Limit', 'Num_Credit_Inquiries', 'Outstanding_Debt', 'Credit_Utilization_Ratio',
                   'Credit_History_Age', 'Total_EMI_per_month', 'Amount_invested_monthly', 'Monthly_Balance']

# Define the appropriate scalers for each column based on the columns characteristics
scalers = {
    'Age': MinMaxScaler(),
    'Annual_Income': RobustScaler(),
    'Monthly_Inhand_Salary': RobustScaler(),
    'Num_Bank_Accounts': MinMaxScaler(),
    'Num_Credit_Card': MinMaxScaler(),
    'Interest_Rate': MinMaxScaler(),
    'Num_of_Loan': MinMaxScaler(),
    'Delay_from_due_date': RobustScaler(),
    'Num_of_Delayed_Payment': RobustScaler(),
    'Changed_Credit_Limit': RobustScaler(),
    'Num_Credit_Inquiries': MinMaxScaler(),
    'Outstanding_Debt': RobustScaler(),
    'Credit_Utilization_Ratio': MinMaxScaler(),
    'Credit_History_Age': MinMaxScaler(),
    'Total_EMI_per_month': MinMaxScaler(),
    'Amount_invested_monthly': MinMaxScaler(),
    'Monthly_Balance': MinMaxScaler()
}

# Function to scale a single column
def scale_column(scaler, column, X_train, X_test):
    X_train[column] = scaler.fit_transform(X_train[[column]])
    X_test[column] = scaler.transform(X_test[[column]])

# Incremental scaling function so that we hopefully do not run out of memory anymore
def incremental_scaling(scalers, columns, X_train, X_test, batch_size):
    for i in range(0, len(columns), batch_size):
        batch_columns = columns[i:i + batch_size]
        for column in batch_columns:
            scaler = scalers[column]
            scale_column(scaler, column, X_train, X_test)

# Set batch size to something small to help with memory
batch_size = 5

# Perform incremental scaling
incremental_scaling(scalers, numeric_columns, X_train, X_test, batch_size)


# To show that the dataset has been preprocessed
X_train.head(10)

X_train.columns

X_test.head(10)

X_test.columns

y_train.head(10)

y_test.head(10)

# Save X_train, X_test, y_train, and y_test to a pickle file so I can reference them in the model building part

# Do the importing 
import pickle

# Save the data
with open('train_test_data.pkl', 'wb') as f:
    pickle.dump((X_train, X_test, y_train, y_test), f)
