==> Notebooks/4 Model Building.ipynb <==
# Model Building Section

import pickle

# Load the data from the pickle file
with open('train_test_data.pkl', 'rb') as f:
    X_train, X_test, y_train, y_test = pickle.load(f)


from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.neural_network import MLPClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.gaussian_process import GaussianProcessClassifier
from sklearn.ensemble import BaggingClassifier, AdaBoostClassifier
from sklearn.ensemble import StackingClassifier

# Excluding the unncessary columns from the input features for the models by dropping them now
columns_to_exclude = ['ID', 'Customer_ID', 'Month', 'SSN', 'Type_of_Loan', 'Name']  # List of columns to exclude
X_train = X_train.drop(columns=columns_to_exclude)
X_test = X_test.drop(columns=columns_to_exclude)

# XG Boost

# !pip install xgboost
# Ran this code one to install, now it is being commented out

from sklearn.preprocessing import LabelEncoder
import xgboost as xgb

# Initialize LabelEncoder
label_encoder = LabelEncoder()

# Encode the target variable
y_train_encoded = label_encoder.fit_transform(y_train)
y_test_encoded = label_encoder.transform(y_test)

# Instantiate XGBoost classifier
xgb_classifier = xgb.XGBClassifier()

# Train XGBoost classifier using the encoded target variables
xgb_classifier.fit(X_train, y_train_encoded)

# Predict
y_pred_xgb = xgb_classifier.predict(X_test)

# Evaluate accuracy
accuracy_xgb = accuracy_score(y_test_encoded, y_pred_xgb)
print("XGBoost Classifier Accuracy:", accuracy_xgb)

## Bagging

# Ensemble Methods - Bagging
bagging = BaggingClassifier()
bagging.fit(X_train, y_train)
y_pred_bagging = bagging.predict(X_test)
bagging_score = accuracy_score(y_test, y_pred_bagging)
print("Bagging Classifier Accuracy:", bagging_score)

## Logistic Regression

# Train the logistic regression model
logistic_regression_model = LogisticRegression()
logistic_regression_model.fit(X_train, y_train)

# Predictions using logistic regression model
logistic_regression_preds = logistic_regression_model.predict(X_test)

# Evaluate the logistic regression model
logistic_regression_accuracy = logistic_regression_model.score(X_test, y_test)
print("Logistic Regression Accuracy:", logistic_regression_accuracy)

# All of the other models follow the same basic procedures

## Random Forest Classifier

from sklearn.ensemble import RandomForestClassifier

rf = RandomForestClassifier(criterion = 'entropy', n_estimators = 500, max_depth = 5, random_state = 101)

rf.fit(X_train, y_train)

y_pred_rf = rf.predict(X_test)

rf_score = accuracy_score(y_test, y_pred_rf)

print("Random Forest Classifier Accuracy:", rf_score)

## Gradient Boosting

# Gradient Boosting Machine 
gbm = GradientBoostingClassifier()
gbm.fit(X_train, y_train)
y_pred_gbm = gbm.predict(X_test)
gbm_score = accuracy_score(y_test, y_pred_gbm)
print("GBM Accuracy:", gbm_score)

## Naive Bayes

# Naive Bayes
nb = GaussianNB()
nb.fit(X_train, y_train)
y_pred_nb = nb.predict(X_test)
nb_score = accuracy_score(y_test, y_pred_nb)
print("Naive Bayes Accuracy:", nb_score)

## Neural Network

# Neural Network
mlp = MLPClassifier()
mlp.fit(X_train, y_train)
y_pred_mlp = mlp.predict(X_test)
mlp_score = accuracy_score(y_test, y_pred_mlp)
print("Neural Network Accuracy:", mlp_score)

## Decision Tree

# Decision Tree
dt = DecisionTreeClassifier()
dt.fit(X_train, y_train)
y_pred_dt = dt.predict(X_test)
dt_score = accuracy_score(y_test, y_pred_dt)
print("Decision Tree Accuracy:", dt_score)

## AdaBoost

# Ensemble Methods - AdaBoost
adaboost = AdaBoostClassifier()
adaboost.fit(X_train, y_train)
y_pred_adaboost = adaboost.predict(X_test)
adaboost_score = accuracy_score(y_test, y_pred_adaboost)
print("AdaBoost Classifier Accuracy:", adaboost_score)

## Support Vector Machine (SVM)

# This one takes a while to run
svm = SVC()
svm.fit(X_train, y_train)
y_pred_svm = svm.predict(X_test)
svm_score = accuracy_score(y_test, y_pred_svm)
print("SVM Accuracy:", svm_score)

## Stacking

# Ensemble Methods - Stacking (running this one takes a long time)
estimators = [('dt', DecisionTreeClassifier()), ('bagging', BaggingClassifier()), ('gbm', GradientBoostingClassifier())]
stacking = StackingClassifier(estimators=estimators)
stacking.fit(X_train, y_train)
y_pred_stacking = stacking.predict(X_test)
stacking_score = accuracy_score(y_test, y_pred_stacking)
print("Stacking Classifier Accuracy:", stacking_score)

import pandas as pd

# Define evaluation data
evaluation_data = [("XG Boost", accuracy_xgb), ("Bagging", bagging_score), 
                   ("Random Forest", rf_score), ("Gradient Boosting", gbm_score),
                   ("Naive Bayes", nb_score),
                   ("Neural Network", mlp_score), ("Decision Tree", dt_score),
                   ("Adaboost", adaboost_score), ("SVM", svm_score), ("Stacking", stacking_score)]

# Create a pandas DataFrame
evaluation_df = pd.DataFrame(evaluation_data, columns=["Model Name", "Accuracy Score"])

# Sort DataFrame by 'R2 Score' column in descending order
evaluation_df = evaluation_df.sort_values(by="Accuracy Score", ascending=False)

# Display the DataFrame
print(evaluation_df)


Although the Stacking ensemble method had a slightly higher accuracy, we will use Bagging as our final model because the computation time is significantly faster, and the accuracy is only about 0.003 worse than the stacking model. 

## Classification Reports

# Print the classification reports of each of the models

from sklearn.metrics import classification_report

# Transform predicted labels back to original classes just for XG Boost
y_pred_xgb_original = label_encoder.inverse_transform(y_pred_xgb)

print("XG Boost:")
print(classification_report(y_test, y_pred_xgb_original))

print("Bagging:")
print(classification_report(y_test, y_pred_bagging))

print("Logistic Regression:")
print(classification_report(y_test, logistic_regression_preds))

print("Random Forest:")
print(classification_report(y_test, y_pred_rf))

print("Gradient Boosting:")
print(classification_report(y_test, y_pred_gbm))

print("Naive Bayes:")
print(classification_report(y_test, y_pred_nb))

print("Neural Network:")
print(classification_report(y_test, y_pred_mlp))

print("Decision Tree:")
print(classification_report(y_test, y_pred_dt))

print("Adaboost:")
print(classification_report(y_test, y_pred_adaboost))

print("SVM:")
print(classification_report(y_test, y_pred_svm))

print("Stacking:")
print(classification_report(y_test, y_pred_stacking))

The F1 scores generated in the classification report also show that XG Boost and Bagging are the best models. Based on both the accuracy scores and the classification reports, both the XG Boost model and the Bagging model will move on to the parameter tuning and complex model design stage.
